+++ Coimbra.Services.Timers/TimerSystem.cs
@@ -85,9 +85,10 @@ namespace Coimbra.Services.Timers
             _timerComponentPool.Push(context);
         }
 
-        protected override void OnDestroyed()
+        /// <inheritdoc/>
+        protected override void OnDispose()
         {
-            base.OnDestroyed();
+            base.OnDispose();
             StopAllTimers();
             _timerComponentPool.Initialize(0, 0);
             _timerComponentPool = null;